--- resources/views/auth/auth.blade.php
@@ -1,101 +1,0 @@
-<!doctype html>
-<html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-
-    <!-- CSRF Token -->
-    <meta name="csrf-token" content="{{ csrf_token() }}">
-
-
-    <!-- Scripts -->
-    <script src="https://code.jquery.com/jquery-3.6.0.min.js"
-            integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js"
-            integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN"
-            crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.6.0/dist/js/bootstrap.min.js"
-            integrity="sha384-+YQ4JLhjyBLPDQt//I+STsc9iw4uQqACwlvpslubQzn4u2UU2UFM80nGisd026JF"
-            crossorigin="anonymous"></script>
-    <script src="https://cdn.jsdelivr.net/npm/admin-lte@3.1/dist/js/adminlte.min.js"></script>
-    <script src="https://unpkg.com/alpinejs@3.10.2/dist/cdn.min.js" defer></script>
-
-    <!-- Fonts -->
-    <link rel="dns-prefetch" href="//fonts.gstatic.com">
-    <link rel="stylesheet"
-          href="https://fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,400i,700&amp;display=fallback">
-    {{-- <link href="https://fonts.googleapis.com/css?family=Nunito" rel="stylesheet"> --}}
-
-    <!-- Styles -->
-    {{-- <link href="{{ asset('css/app.css') }}" rel="stylesheet"> --}}
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/admin-lte@3.1/dist/css/adminlte.min.css">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/notyf@3/notyf.min.css">
-
-    <!-- Icons -->
-    <link href="https://unpkg.com/ionicons@4.5.10-0/dist/css/ionicons.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.min.css" integrity="sha512-xh6O/CkQoPOWDdYTDqeRdPCVd1SpvCA9XXcUnZS2FmJNp1coAFzvtCN9BmamE+4aHK8yyUHUSCcJHgXloTyT2A==" crossorigin="anonymous" referrerpolicy="no-referrer" />
-
-
-    <!-- Small Ionicons Fixes for AdminLTE -->
-    <style>
-        .minh-100 {
-            height: 70vh;
-        }
-
-        .nav-icon.icon:before {
-            width: 25px;
-        }
-    </style>
-
-
-    <script type="module"></script>
-</head>
-
-<body class="layout-fixed">
-<div id="app">
-    <main class="p-5 ">
-        @yield('content')
-    </main>
-</div>
-
-@stack('modals')
-
-<script src="https://cdn.jsdelivr.net/gh/livewire/turbolinks@v0.1.x/dist/livewire-turbolinks.js"
-        data-turbolinks-eval="false" data-turbo-eval="false"></script>
-
-@stack('scripts')
-
-<script src="https://cdn.jsdelivr.net/npm/notyf@3/notyf.min.js"></script>
-
-@if (session()->has('success'))
-    <script>
-        const notyf = new Notyf({dismissible: true})
-        notyf.success('{{ session('success') }}')
-    </script>
-@endif
-
-<script>
-    /* Simple Alpine Image Viewer */
-    document.addEventListener('alpine:init', () => {
-        Alpine.data('imageViewer', (src = '') => {
-            return {
-                imageUrl: src,
-                refreshUrl() {
-                    this.imageUrl = this.$el.getAttribute("image-url")
-                },
-                fileChosen(event) {
-                    this.fileToDataUrl(event, src => this.imageUrl = src)
-                },
-                fileToDataUrl(event, callback) {
-                    if (!event.target.files.length) return
-                    let file = event.target.files[0],
-                        reader = new FileReader()
-                    reader.readAsDataURL(file)
-                    reader.onload = e => callback(e.target.result)
-                },
-            }
-        })
-    })
-</script>
-</body>
-</html>
